requirements.txt
pytest
playwright
pytest-playwright
boto3
python-dotenv
allure-pytest
logzero
jsonschema
pytest-xdist